Fixed schedule bug. Introduced new GameViewModel and refactored ScheduleViewModel.
